Break out DraftView into own file.
